tests/wpt/tests/css/css-backgrounds/parsing/background-origin-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Backgrounds and Borders Module Level 3: parsing background-origin with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-backgrounds/#background-origin">
<meta name="assert" content="background-origin supports the full grammar '<box>#'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("background-origin", "border-box");
test_valid_value("background-origin", "padding-box");
test_valid_value("background-origin", "content-box");

test_valid_value("background-origin", "border-box, padding-box, content-box");
</script>
</body>
</html>
